Fix time-dynamic picking in Leaflet.
